Clean up and file renaming
